+++ Geofences/GFAppDelegate.m
@@ -21,7 +21,8 @@
 #endif
     
     //Register the app id of the application you created on https://app.contexthub.com
-    [ContextHub registerWithAppId:@"YOUR-GEOFENCE-APP-ID-HERE"];
+    //[ContextHub registerWithAppId:@"YOUR-GEOFENCE-APP-ID-HERE"];
+    [ContextHub registerWithAppId:@"4e0aad2a-b052-42e0-93ee-6f024d11de10"];
     
     //Set the app delegate as the Datasource and Delegate of the Sensor Pipeline so that we can tap into the events.
     [[CCHSensorPipeline sharedInstance] setDelegate:self];
@@ -40,26 +41,26 @@
 - (BOOL)sensorPipeline:(CCHSensorPipeline *)sensorPipeline shouldPostEvent:(NSDictionary *)event {
     // If you'd like to keep events from hitting the server, you can return NO here.
     // This is a good spot to filter events.
-    NSLog(@"GH: Should post event %@", event);
+    NSLog(@"GF: Should post event %@", event);
 
     return YES;
 }
 
 - (void)sensorPipeline:(CCHSensorPipeline *)sensorPipeline willPostEvent:(NSDictionary *)event {
     // If you want to access event data directl before it will be posted to the server, you can do that here
-    NSLog(@"GH: Will post event %@", event);
+    NSLog(@"GF: Will post event %@", event);
 }
 
 - (void)sensorPipeline:(CCHSensorPipeline *)sensorPipeline didPostEvent:(NSDictionary *)event {
     // If you want to access event data directly after it has been posted to the server, you can do that here
-    NSLog(@"GH: Did post event %@", event);
+    NSLog(@"GF: Did post event %@", event);
 }
 
 #pragma mark - Sensor Pipeline Data Source
 
 - (NSDictionary *)sensorPipeline:(CCHSensorPipeline *)sensorPipeline payloadForEvent:(NSDictionary *)event {
     // Add custom data structures to the events, and they will end up on the server in the payload property
-    return @{@"custom":@"data"};
+    return @{};
 }
 
 @end